+++ aksel.nav.no/website/pages/eksempler/button/nextjs-link.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 
 const Example = () => {
   return (
-    <Link href="#" passHref>
+    <Link href="#" passHref legacyBehavior>
       <Button as="a">Lenke</Button>
     </Link>
   );