+++ @navikt/internal/css/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @navikt/ds-css-internal
 
+## 2.4.0
+
 ## 2.3.1
 
 ### Patch Changes